+++ src/Phan/Bootstrap.php
@@ -35,7 +35,7 @@ if (function_exists('uopz_allow_exit') && !ini_get('uopz.disable')) {
 if (PHP_VERSION_ID < 70200) {
     fprintf(
         STDERR,
-        "ERROR: Phan 4.x requires PHP 7.2+ to run, but PHP %s is installed." . PHP_EOL,
+        "ERROR: Phan 5.x requires PHP 7.2+ to run, but PHP %s is installed." . PHP_EOL,
         PHP_VERSION
     );
     fwrite(STDERR, "PHP 7.1 reached its end of life in December 2019." . PHP_EOL);
@@ -44,7 +44,7 @@ if (PHP_VERSION_ID < 70200) {
     exit(1);
 }
 
-const LATEST_KNOWN_PHP_AST_VERSION = '1.0.13';
+const LATEST_KNOWN_PHP_AST_VERSION = '1.0.14';
 
 /**
  * Dump instructions on how to install php-ast
@@ -61,9 +61,9 @@ function phan_output_ast_installation_instructions(): void
         $extension_dir .= ' (extension directory does not exist and may need to be changed)';
     }
     if (DIRECTORY_SEPARATOR === '\\') {
-        if (PHP_VERSION_ID < 80100 || !preg_match('/[a-zA-Z]/', PHP_VERSION)) {
-            // e.g. https://windows.php.net/downloads/pecl/releases/ast/1.0.10/php_ast-1.0.10-8.0-nts-vs16-x64.zip for php 8.0, 64-bit non thread safe
-            // e.g. https://windows.php.net/downloads/pecl/releases/ast/1.0.10/php_ast-1.0.10-7.4-ts-vc15-x86.zip for php 7.4, 32-bit thread safe
+        if (PHP_VERSION_ID >= 70300 && PHP_VERSION_ID < 80100 || !preg_match('/[a-zA-Z]/', PHP_VERSION)) {
+            // e.g. https://windows.php.net/downloads/pecl/releases/ast/1.0.14/php_ast-1.0.14-8.0-nts-vs16-x64.zip for php 8.0, 64-bit non thread safe
+            // e.g. https://windows.php.net/downloads/pecl/releases/ast/1.0.14/php_ast-1.0.14-7.4-ts-vc15-x86.zip for php 7.4, 32-bit thread safe
             fprintf(
                 STDERR,
                 PHP_EOL . "Windows users can download php-ast from https://windows.php.net/downloads/pecl/releases/ast/%s/php_ast-%s-%s-%s-%s-%s.zip" . PHP_EOL,
@@ -76,12 +76,19 @@ function phan_output_ast_installation_instructions(): void
             );
             fwrite(STDERR, "(if that link doesn't work, check https://windows.php.net/downloads/pecl/releases/ast/ )" . PHP_EOL);
             fwrite(STDERR, "To install php-ast, add php_ast.dll from the zip to $extension_dir," . PHP_EOL);
-            fwrite(STDERR, "Then, enable php-ast by adding the following lines to your php.ini file at '$ini_path'" . PHP_EOL . PHP_EOL);
-            if (!is_dir((string)$configured_extension_dir) && is_dir($new_extension_dir)) {
-                fwrite(STDERR, "extension_dir=$new_extension_dir" . PHP_EOL);
+        } else {
+            if (PHP_VERSION_ID < 70300) {
+                fwrite(STDERR, "php-ast 1.0.11 is the minimum php-ast version needed for ast version 85. https://pecl.php.net/package/ast/1.0.11/windows does not supply dlls for php 7.2 because php-ast 1.0.11 was published after security support for php 7.2 was dropped" . PHP_EOL);
+            } else {
+                fprintf(STDERR, "Releases for php %s may not yet be available at https://windows.php.net/downloads/pecl/releases/ast/" . PHP_EOL, PHP_VERSION);
             }
-            fwrite(STDERR, "extension=php_ast.dll" . PHP_EOL . PHP_EOL);
+            fwrite(STDERR, "To build php-ast from source for Windows, see https://wiki.php.net/internals/windows/stepbystepbuild_sdk_2 and https://wiki.php.net/internals/windows/stepbystepbuild_sdk_2#building_pecl_extensions" . PHP_EOL);
         }
+        fwrite(STDERR, "Then, enable php-ast by adding the following lines to your php.ini file at '$ini_path'" . PHP_EOL . PHP_EOL);
+        if (!is_dir((string)$configured_extension_dir) && is_dir($new_extension_dir)) {
+            fwrite(STDERR, "extension_dir=$new_extension_dir" . PHP_EOL);
+        }
+        fwrite(STDERR, "extension=php_ast.dll" . PHP_EOL . PHP_EOL);
     } else {
         fwrite(STDERR, <<<EOT
 php-ast can be installed in the following ways:
@@ -127,7 +134,7 @@ if (extension_loaded('ast')) {
     }
     $phan_output_ast_too_old_and_exit = /** @return never */ static function (string $minimum_ast_version, string $php_version_bound) use ($ast_version): void {
         $error_message = sprintf(
-            "Phan 4.x requires php-ast %s+ to properly analyze ASTs for php %s+. php-ast %s and php %s is installed." . PHP_EOL,
+            "Phan 5.x requires php-ast %s+ to properly analyze ASTs for php %s+. php-ast %s and php %s is installed." . PHP_EOL,
             $minimum_ast_version,
             $php_version_bound,
             $ast_version,
@@ -139,21 +146,21 @@ if (extension_loaded('ast')) {
         exit(1);
     };
 
-    if (PHP_VERSION_ID >= 80100 && version_compare($ast_version, '1.0.13') < 0) {
-        $phan_output_ast_too_old_and_exit('1.0.13', '8.1');
-    } elseif (PHP_VERSION_ID >= 80000 && version_compare($ast_version, '1.0.10') < 0) {
-        $phan_output_ast_too_old_and_exit('1.0.10', '8.0');
+    if (PHP_VERSION_ID >= 80100 && version_compare($ast_version, '1.0.14') < 0) {
+        $phan_output_ast_too_old_and_exit('1.0.14', '8.1');
+    } elseif (PHP_VERSION_ID >= 80000 && version_compare($ast_version, '1.0.11') < 0) {
+        $phan_output_ast_too_old_and_exit('1.0.11', '8.0');
     } elseif (PHP_VERSION_ID >= 70400 && version_compare($ast_version, '1.0.2') < 0) {
         fprintf(
             STDERR,
-            "WARNING: Phan 4.x requires php-ast 1.0.2+ to properly analyze ASTs for php 7.4+ (1.0.10+ is recommended). php-ast %s and php %s is installed." . PHP_EOL,
+            "WARNING: Phan 5.x requires php-ast 1.0.2+ to properly analyze ASTs for php 7.4+ (1.0.14+ is recommended). php-ast %s and php %s is installed." . PHP_EOL,
             $ast_version,
             PHP_VERSION
         );
         phan_output_ast_installation_instructions();
     } elseif (version_compare($ast_version, '1.0.0') <= 0) {
         $error_message = sprintf(
-            "Phan 4.x requires php-ast %s+ because it depends on AST version %d. php-ast '%s' is installed." . PHP_EOL,
+            "Phan 5.x requires php-ast %s+ because it depends on AST version %d. php-ast '%s' is installed." . PHP_EOL,
             Config::MINIMUM_AST_EXTENSION_VERSION,
             Config::AST_VERSION,
             $ast_version
@@ -164,8 +171,10 @@ if (extension_loaded('ast')) {
         exit(1);
     }
     // @phan-suppress-next-line PhanRedundantCondition, PhanImpossibleCondition, PhanSuspiciousValueComparison
-    if (PHP_VERSION_ID >= 80000 && version_compare(PHP_VERSION, '8.0.0') < 0) {
-        fwrite(STDERR, "WARNING: Phan may not work properly in PHP 8 versions before PHP 8.0.0. The currently used PHP version is " . PHP_VERSION . PHP_EOL);
+    if (PHP_VERSION_ID < 80100 && PHP_VERSION_ID % 100 === 0 && PHP_EXTRA_VERSION !== '') {
+        // Warn for 8.0.0RC1, 7.4.0alpha1, 7.3.0-dev, etc.
+        // But don't warn for 8.1.0 since there's no way to upgrade to a stable release.
+        fwrite(STDERR, "WARNING: Phan may not work properly in versions prior to the first stable release of a php minor version. The currently used PHP version is " . PHP_VERSION . PHP_EOL);
     }
     unset($ast_version);
 }